backend/requirements.txt
fastapi
uvicorn
python-jose
python-multipart==0.0.6
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0 